+++ csatar-plugins/knowledgerepository/lang/en/lang.php
@@ -15,7 +15,8 @@
                     'workPlans' => 'Work Plans',
                     'methodologies' => 'Methodologies',
                     'songTypes' => 'Song types',
-                    'folkSongTypes' => 'Folk song types'
+                    'folkSongTypes' => 'Folk song types',
+                    'regions' => 'Regions'
                 ],
                 'knowledgeRepositoryParameters' => [
                     'knowledgeRepositoryParameters' => 'Knowledge Repository Parameters',
@@ -39,7 +40,11 @@
                     'methodology' => 'Methodology',
                     'ageGroup' => 'Age group',
                     'songType' => 'Song type',
-                    'folkSongType' => 'Folk song type'
+                    'folkSongType' => 'Folk song type',
+                    'region' => 'Region',
+                    'bigRegion' => 'Big Region',
+                    'midRegion' => 'Mid Region',
+                    'smallRegion' => 'Small Region',
                 ],
             ],
             'general' => [
@@ -60,6 +65,7 @@
                 'import' => 'Import',
                 'row' => 'row',
                 'file' => 'File',
+                'select' => 'Select...'
             ],
             'game' => [
                 'game' => 'Game',
@@ -120,7 +126,7 @@
                 'cannotFindTrialSystemTrialType' => 'Cannot find trial system trial type(s): ',
                 'errorsOccurred' => 'Errors occurred during import: ',
                 'importSuccessful' => 'Import successful!',
-            ],
+        ],
             'workPlan' => [
                 'workPlan' => 'Work Plan',
                 'workPlans' => 'Work Plans',