+++ csatar-plugins/knowledgerepository/updates/SeederData.php
@@ -4,6 +4,7 @@ use Csatar\KnowledgeRepository\Models\AccidentRiskLevel;
 use Csatar\KnowledgeRepository\Models\FolkSongType;
 use Csatar\KnowledgeRepository\Models\GameDevelopmentGoal;
 use Csatar\KnowledgeRepository\Models\MethodologyType;
+use Csatar\KnowledgeRepository\Models\Region;
 use Csatar\KnowledgeRepository\Models\SongType;
 use Csatar\KnowledgeRepository\Models\Tool;
 use Csatar\KnowledgeRepository\Models\Headcount;
@@ -525,6 +526,237 @@ class SeederData extends Seeder
                 'name' => 'Arató- és summásdalok',
                 'description' => 'a mezei munkásság dalai'
             ],
+        ],
+        'regions' => [
+            [
+                'name' => 'Alföld'
+            ],
+            [
+                'name' => 'Bácska',
+                'big_parent' => 'Alföld'
+            ],
+            [
+                'name' => 'Kis- és Nagykunság',
+                'big_parent' => 'Alföld'
+            ],
+            [
+                'name' => 'Jászság',
+                'big_parent' => 'Alföld'
+            ],
+            [
+                'name' => 'Hajdúság',
+                'big_parent' => 'Alföld'
+            ],
+            [
+                'name' => 'Nyírség',
+                'big_parent' => 'Alföld'
+            ],
+            [
+                'name' => 'Érmellék',
+                'big_parent' => 'Alföld'
+            ],
+            [
+                'name' => 'Szamoshát',
+                'big_parent' => 'Alföld'
+            ],
+            [
+                'name' => 'Tiszahát',
+                'big_parent' => 'Alföld'
+            ],
+            [
+                'name' => 'Dunántúl'
+            ],
+            [
+                'name' => 'Ormánság',
+                'big_parent' => 'Dunántúl'
+            ],
+            [
+                'name' => 'Drávaszög',
+                'big_parent' => 'Dunántúl'
+            ],
+            [
+                'name' => 'Szlavóniai magyarság',
+                'big_parent' => 'Dunántúl'
+            ],
+            [
+                'name' => 'Őrség',
+                'big_parent' => 'Dunántúl'
+            ],
+            [
+                'name' => 'Göcsej',
+                'big_parent' => 'Dunántúl'
+            ],
+            [
+                'name' => 'Bakony erdeje',
+                'big_parent' => 'Dunántúl'
+            ],
+            [
+                'name' => 'Balatonfelvidék',
+                'big_parent' => 'Dunántúl'
+            ],
+            [
+                'name' => 'Kisalföld',
+                'big_parent' => 'Dunántúl'
+            ],
+            [
+                'name' => 'Csallóköz',
+                'big_parent' => 'Dunántúl'
+            ],
+            [
+                'name' => 'Felföld/Felvidék',
+            ],
+            [
+                'name' => 'Palócföld',
+                'big_parent' => 'Felföld/Felvidék'
+            ],
+            [
+                'name' => 'Zobor vidéke',
+                'big_parent' => 'Felföld/Felvidék'
+            ],
+            [
+                'name' => 'Garam mente',
+                'big_parent' => 'Felföld/Felvidék'
+            ],
+            [
+                'name' => 'Hegyalja',
+                'big_parent' => 'Felföld/Felvidék'
+            ],
+            [
+                'name' => 'Erdély',
+            ],
+            [
+                'name' => 'Partium',
+                'big_parent' => 'Erdély'
+            ],
+            [
+                'name' => 'Máramarosi',
+                'big_parent' => 'Erdély',
+                'mid_parent' => 'Partium',
+            ],
+            [
+                'name' => 'Szilágysági',
+                'big_parent' => 'Erdély',
+                'mid_parent' => 'Partium',
+            ],
+            [
+                'name' => 'Szatmári',
+                'big_parent' => 'Erdély',
+                'mid_parent' => 'Partium',
+            ],
+            [
+                'name' => 'Bihari',
+                'big_parent' => 'Erdély',
+                'mid_parent' => 'Partium',
+            ],
+            [
+                'name' => 'Érmelléki',
+                'big_parent' => 'Erdély',
+                'mid_parent' => 'Partium',
+            ],
+            [
+                'name' => 'Kalotaszeg',
+                'big_parent' => 'Erdély'
+            ],
+            [
+                'name' => 'Torockó',
+                'big_parent' => 'Erdély'
+            ],
+            [
+                'name' => 'Mezőség',
+                'big_parent' => 'Erdély'
+            ],
+            [
+                'name' => 'Székelyföld',
+                'big_parent' => 'Erdély'
+            ],
+            [
+                'name' => 'Aranyosszék',
+                'big_parent' => 'Erdély',
+                'mid_parent' => 'Székelyföld',
+            ],
+            [
+                'name' => 'Marosszék',
+                'big_parent' => 'Erdély',
+                'mid_parent' => 'Székelyföld',
+            ],
+            [
+                'name' => 'Udvarhelyszék',
+                'big_parent' => 'Erdély',
+                'mid_parent' => 'Székelyföld',
+            ],
+            [
+                'name' => 'Háromszék',
+                'big_parent' => 'Erdély',
+                'mid_parent' => 'Székelyföld',
+            ],
+            [
+                'name' => 'Kézdiszék',
+                'big_parent' => 'Erdély',
+                'mid_parent' => 'Székelyföld',
+                'small_parent' => 'Háromszék',
+            ],
+            [
+                'name' => 'Orbaiszék',
+                'big_parent' => 'Erdély',
+                'mid_parent' => 'Székelyföld',
+                'small_parent' => 'Háromszék',
+            ],
+            [
+                'name' => 'Sepsiszék',
+                'big_parent' => 'Erdély',
+                'mid_parent' => 'Székelyföld',
+                'small_parent' => 'Háromszék',
+            ],
+            [
+                'name' => 'Csíkszék',
+                'big_parent' => 'Erdély',
+                'mid_parent' => 'Székelyföld',
+            ],
+            [
+                'name' => 'Alcsík',
+                'big_parent' => 'Erdély',
+                'mid_parent' => 'Székelyföld',
+                'small_parent' => 'Csíkszék',
+            ],
+            [
+                'name' => 'Felcsík',
+                'big_parent' => 'Erdély',
+                'mid_parent' => 'Székelyföld',
+                'small_parent' => 'Csíkszék',
+            ],
+            [
+                'name' => 'Gyergyószék',
+                'big_parent' => 'Erdély',
+                'mid_parent' => 'Székelyföld',
+            ],
+            [
+                'name' => 'Kászon',
+                'big_parent' => 'Erdély',
+                'mid_parent' => 'Székelyföld',
+            ],
+            [
+                'name' => 'Barcaság',
+                'big_parent' => 'Erdély',
+                'mid_parent' => 'Székelyföld',
+            ],
+            [
+                'name' => 'Gyímes',
+                'big_parent' => 'Erdély',
+                'mid_parent' => 'Székelyföld',
+            ],
+            [
+                'name' => 'Moldva',
+                'big_parent' => 'Erdély',
+                'mid_parent' => 'Székelyföld',
+            ],
+            [
+                'name' => 'Bukovina',
+                'big_parent' => 'Erdély',
+                'mid_parent' => 'Székelyföld',
+            ],
+            [
+                'name' => 'Nem alkalmazható',
+            ],
         ]
     ];
 
@@ -667,6 +899,30 @@ class SeederData extends Seeder
             ]);
             $folkSongType->description = $folkSongTypeData['description'];
             $folkSongType->save();
+        }
+
+        // Region type
+        foreach ($this::DATA['regions'] as $regionData) {
+            $region = Region::firstOrNew([
+                'name' => $regionData['name']
+            ]);
+
+            if (isset($regionData['big_parent'])) {
+                $parent = Region::where('name', $regionData['big_parent'])->first() ?? null;
+                $region->big_parent_id = $parent->id;
+            }
+
+            if (isset($regionData['mid_parent'])) {
+                $parent = Region::where('name', $regionData['mid_parent'])->first() ?? null;
+                $region->mid_parent_id = $parent->id;
+            }
+
+            if (isset($regionData['small_parent'])) {
+                $parent = Region::where('name', $regionData['small_parent'])->first() ?? null;
+                $region->small_parent_id = $parent->id;
+            }
+
+            $region->save();
         }
     }
 }