+++ Changelog.md
@@ -14,6 +14,7 @@
 - CS-475 - Added option to backend scout list to (soft)delete scout with personal data
 - CS-516 - Created Song type model
 - CS-517 - Created Folk song type model
+- CS-518 - Created Region model
 
 ## 1.11.0
 ### 2023-04-13